refactor: rename Geometry folder to geometry
