+++ source/dynamical_systems/include/dynamical_systems/Circular.hpp
@@ -6,7 +6,7 @@
 #pragma once
 
 #include "dynamical_systems/DynamicalSystem.hpp"
-#include "state_representation/Geometry/Ellipsoid.hpp"
+#include "state_representation/geometry/Ellipsoid.hpp"
 #include "state_representation/Parameters/Parameter.hpp"
 #include "state_representation/Space/Cartesian/CartesianPose.hpp"
 #include "state_representation/Space/Cartesian/CartesianState.hpp"
